Move secrets out of internal/
