+++ internal/admin/config.go
@@ -23,10 +23,10 @@ import (
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/internal/database"
-	"github.com/google/exposure-notifications-server/internal/secrets"
 	"github.com/google/exposure-notifications-server/internal/setup"
 	"github.com/google/exposure-notifications-server/internal/signing"
 	"github.com/google/exposure-notifications-server/internal/storage"
+	"github.com/google/exposure-notifications-server/pkg/secrets"
 )
 
 var _ setup.BlobstoreConfigProvider = (*Config)(nil)
